+++ cmd/rp-indexer/main.go
@@ -12,8 +12,10 @@ import (
 	"github.com/getsentry/sentry-go"
 	_ "github.com/lib/pq"
 	"github.com/nyaruka/ezconf"
+	"github.com/nyaruka/gocommon/aws/cwatch"
 	indexer "github.com/nyaruka/rp-indexer/v9"
 	"github.com/nyaruka/rp-indexer/v9/indexers"
+	"github.com/nyaruka/rp-indexer/v9/runtime"
 	slogmulti "github.com/samber/slog-multi"
 	slogsentry "github.com/samber/slog-sentry"
 )
@@ -25,7 +27,7 @@ var (
 )
 
 func main() {
-	cfg := indexer.NewDefaultConfig()
+	cfg := runtime.NewDefaultConfig()
 	loader := ezconf.NewLoader(cfg, "indexer", "Indexes RapidPro contacts to ElasticSearch", []string{"indexer.toml"})
 	loader.MustLoad()
 
@@ -36,15 +38,14 @@ func main() {
 		os.Exit(1)
 	}
 
+	rt := &runtime.Runtime{Config: cfg}
+
 	// configure our logger
 	logHandler := slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: level})
 	slog.SetDefault(slog.New(logHandler))
 
-	logger := slog.With("comp", "main")
-	logger.Info("starting indexer", "version", version, "released", date)
-
 	// if we have a DSN entry, try to initialize it
-	if cfg.SentryDSN != "" {
+	if rt.Config.SentryDSN != "" {
 		err := sentry.Init(sentry.ClientOptions{
 			Dsn:           cfg.SentryDSN,
 			EnableTracing: false,
@@ -56,7 +57,7 @@ func main() {
 
 		defer sentry.Flush(2 * time.Second)
 
-		logger = slog.New(
+		logger := slog.New(
 			slogmulti.Fanout(
 				logHandler,
 				slogsentry.Option{Level: slog.LevelError}.NewSentryHandler(),
@@ -66,24 +67,32 @@ func main() {
 		slog.SetDefault(logger)
 	}
 
-	db, err := sql.Open("postgres", cfg.DB)
+	log := slog.With("comp", "main")
+	log.Info("starting indexer", "version", version, "released", date)
+
+	rt.DB, err = sql.Open("postgres", cfg.DB)
 	if err != nil {
-		logger.Error("unable to connect to database")
+		log.Error("unable to connect to database", "error", err)
+	}
+
+	rt.CW, err = cwatch.NewService(rt.Config.AWSAccessKeyID, rt.Config.AWSSecretAccessKey, rt.Config.AWSRegion, rt.Config.CloudwatchNamespace, rt.Config.DeploymentID)
+	if err != nil {
+		log.Error("unable to create cloudwatch service", "error", err)
 	}
 
 	idxrs := []indexers.Indexer{
-		indexers.NewContactIndexer(cfg.ElasticURL, cfg.ContactsIndex, cfg.ContactsShards, cfg.ContactsReplicas, 500),
+		indexers.NewContactIndexer(rt.Config.ElasticURL, rt.Config.ContactsIndex, rt.Config.ContactsShards, rt.Config.ContactsReplicas, 500),
 	}
 
-	if cfg.Rebuild {
+	if rt.Config.Rebuild {
 		// if rebuilding, just do a complete index and quit. In future when we support multiple indexers,
 		// the rebuild argument can be become the name of the index to rebuild, e.g. --rebuild=contacts
 		idxr := idxrs[0]
-		if _, err := idxr.Index(db, true, cfg.Cleanup); err != nil {
-			logger.Error("error during rebuilding", "error", err, "indexer", idxr.Name())
+		if _, err := idxr.Index(rt, true, rt.Config.Cleanup); err != nil {
+			log.Error("error during rebuilding", "error", err, "indexer", idxr.Name())
 		}
 	} else {
-		d := indexer.NewDaemon(cfg, db, idxrs, time.Duration(cfg.Poll)*time.Second)
+		d := indexer.NewDaemon(rt, idxrs)
 		d.Start()
 
 		handleSignals(d)